+++ packages/pigeon/CHANGELOG.md
@@ -3,6 +3,11 @@
 * [swift] [kotlin] Adds support for FFI and JNI.
 * Expose missing PigeonOptions and language-specific configurations as command-line arguments.
 
+## 26.3.4
+
+* [kotlin] Updates generated error class to inherit from `RuntimeException`
+  instead of `Throwable`, for better Java interoperability.
+
 ## 26.3.3
 
 * Updates `analyzer` dependency to support versions 10 through 12.